% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_kernels.R
\name{absolute_dist}
\alias{absolute_dist}
\title{Calculate absolute distance}
\usage{
absolute_dist(X, X_prime, lengthscales = NULL)
}
\arguments{
\item{X}{2-column matrix of longitudes and latitudes}

\item{X_prime}{2-column matrix of longitudes and latitude}

\item{lengthscales}{numeric: to be supplied to \code{get_dist()}}
}
\value{
tensor of distances between all points in \code{X} and all points in \code{X_prime}
}
\description{
Calculate absolute distance
}
\seealso{
\code{\link[=great_circle_dist]{great_circle_dist()}}
}
