% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_kernels.R
\name{degrees_radians}
\alias{degrees_radians}
\alias{degrees_to_radians}
\alias{radians_to_degrees}
\title{Convert degrees to radians or radians to degrees}
\usage{
degrees_to_radians(degrees)

radians_to_degrees(radians)
}
\arguments{
\item{degrees, }{numeric: number of decimal degrees to convert to radians}

\item{radians, }{numeric: number of radians to convert to decimal degrees}
}
\value{
\code{degrees} or \code{radians} (numeric)
}
\description{
Convert degrees to radians or radians to degrees
}
