% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_kernels.R
\name{degrees_to_radians}
\alias{degrees_to_radians}
\title{Convert degrees to radians}
\usage{
degrees_to_radians(degrees)
}
\arguments{
\item{degrees, }{numeric: number of decimal degrees to convert to radians}
}
\value{
\code{degrees} in radians (numeric)
}
\description{
Convert degrees to radians
}
