% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_kernels.R
\name{great_circle_dist}
\alias{great_circle_dist}
\title{Calculate great circle distance}
\usage{
great_circle_dist(X, X_prime, active_dims, circumference = 1L)
}
\arguments{
\item{X}{2-column matrix of longitudes and latitudes}

\item{X_prime}{2-column matrix of longitudes and latitude}

\item{circumference}{numeric: sphere circumference in user-specified units, e.g. Earth: 6378137 metres}

\item{lengthscales}{numeric: to be supplied to \code{get_dist()}}
}
\value{
tensor of distances between all points in \code{X} and all points in \code{X_prime}
}
\description{
Calculate great circle distance
}
\details{
function expects \code{X} and \code{X_prime} to be expressed in radians.
}
\seealso{
\code{\link[=degrees_to_radians]{degrees_to_radians()}}, \code{\link[=circmat]{circmat()}}, \code{\link[=absolute_dist]{absolute_dist()}}
}
